Update get-started-docker.adoc
I am following the training videos which at first two labs, suggest populating "my_distributed_map (underscores)" in SQL and then reading content in Java from "my-distributed-map (dashes)". This is somewhat not easy to spot and arguably incorrect in terms of consistency across the training material. That rather be standardized.